refactor: split out tauri setup code for holochain_service and holochain_bundled into seperate files for clarity
