+++ src-tauri/src/config.rs
@@ -1,0 +1,3 @@
+pub const APP_ID: &'static str = "volla-messages";
+
+pub const HAPP_BUNDLE_BYTES: &'static [u8] = include_bytes!("../../workdir/relay.happ");
